feat: leaderboard API and database restructure
