+++ .env.defaults
@@ -61,6 +61,8 @@ MONGODB_DATABASE='kaetram_devlopment'
 MONGODB_TLS=false
 # Use the `mongodb+srv` syntax
 MONGODB_SRV=false
+# When we're allowed to aggregate new data.
+AGGREGATE_THRESHOLD=60000
 
 # === World Configurations ===
