+++ packages/common/database/mongodb/loader.ts
@@ -1,6 +1,8 @@
 import config from '@kaetram/common/config';
 import log from '@kaetram/common/util/log';
 
+import type Player from '@kaetram/server/src/game/entity/character/player/player';
+import type { Db } from 'mongodb';
 import type { SerializedAbility } from '@kaetram/common/types/ability';
 import type { AchievementData, SerializedAchievement } from '@kaetram/common/types/achievement';
 import type { EquipmentData, SerializedEquipment } from '@kaetram/common/types/equipment';
@@ -8,8 +10,6 @@ import type { QuestData, SerializedQuest } from '@kaetram/common/types/quest';
 import type { SerializedSkills, SkillData } from '@kaetram/common/types/skills';
 import type { SerializedContainer, SlotData } from '@kaetram/common/types/slot';
 import type { StatisticsData } from '@kaetram/common/types/statistics';
-import type { Db } from 'mongodb';
-import type Player from '../../game/entity/character/player/player';
 
 export default class Loader {
     public constructor(private database?: Db) {}
