+++ packages/common/types/leaderboards.d.ts
@@ -1,0 +1,4 @@
+export interface TotalExperience {
+    username: string;
+    totalExperience: number;
+}
