--- packages/hub/src/database/mongodb/creator.ts
@@ -1,5 +1,0 @@
-import type { Db } from 'mongodb';
-
-export default class Creator {
-    public constructor(private database?: Db) {}
-}
