--- packages/hub/src/database/mongodb/loader.ts
@@ -1,5 +1,0 @@
-import type { Db } from 'mongodb';
-
-export default class Loader {
-    public constructor(private database?: Db) {}
-}
