+++ packages/server/src/game/entity/character/player/player.ts
@@ -42,6 +42,16 @@ import { Opcodes, Modules } from '@kaetram/common/network';
 import config from '@kaetram/common/config';
 import { Team } from '@kaetram/common/api/minigame';
 
+import type NPC from '../../npc/npc';
+import type Skill from './skill/skill';
+import type Map from '../../../map/map';
+import type World from '../../../world';
+import type Item from '../../objects/item';
+import type Area from '../../../map/areas/area';
+import type Regions from '../../../map/regions';
+import type Connection from '../../../../network/connection';
+import type Resource from '../../../globals/impl/resource';
+import type Minigame from '../../../minigames/minigame';
 import type { EntityDisplayInfo } from '@kaetram/common/types/entity';
 import type { Bonuses, Stats } from '@kaetram/common/types/item';
 import type { ProcessedDoor } from '@kaetram/common/types/map';
@@ -49,18 +59,8 @@ import type { PlayerData } from '@kaetram/common/types/player';
 import type { PointerData } from '@kaetram/common/types/pointer';
 import type Entities from '@kaetram/server/src/controllers/entities';
 import type Packet from '@kaetram/server/src/network/packet';
-import type { PlayerInfo } from '../../../../database/mongodb/creator';
-import type MongoDB from '../../../../database/mongodb/mongodb';
-import type Connection from '../../../../network/connection';
-import type Resource from '../../../globals/impl/resource';
-import type Area from '../../../map/areas/area';
-import type Map from '../../../map/map';
-import type Regions from '../../../map/regions';
-import type Minigame from '../../../minigames/minigame';
-import type World from '../../../world';
-import type NPC from '../../npc/npc';
-import type Item from '../../objects/item';
-import type Skill from './skill/skill';
+import type { PlayerInfo } from '@kaetram/common/database/mongodb/creator';
+import type MongoDB from '@kaetram/common/database/mongodb/mongodb';
 
 type KillCallback = (character: Character) => void;
 type NPCTalkCallback = (npc: NPC) => void;
